+++ src/simulation/SpeedTargets.cpp
@@ -1,17 +1,17 @@
 #include "simulation/SpeedTargets.hpp"
 
-cda_rail::SpeedTargets::SpeedTargets(std::vector<uint>   timesteps,
+cda_rail::SpeedTargets::SpeedTargets(std::vector<ulong>  timesteps,
                                      std::vector<double> speeds) {
   for (size_t i = 0; i < timesteps.size(); i++) {
     targets.insert({timesteps.at(i), speeds.at(i)});
   };
 };
 
-double cda_rail::SpeedTargets::find_target_speed(uint timestep) {
+double cda_rail::SpeedTargets::find_target_speed(ulong timestep) {
   if (targets.size() == 0)
     throw std::out_of_range("Speed target set is empty.");
 
-  std::map<uint, double>::iterator it = targets.upper_bound(timestep);
+  std::map<ulong, double>::iterator it = targets.upper_bound(timestep);
   if (it != targets.begin()) {
     --it;
   }
